+++ tests/functional/test_cli.py
@@ -33,7 +33,7 @@ def test_run_no_paths(runner, monkeypatch, tmpdir):
 
     result = runner.invoke(cli.main, [])
 
-    assert "actual: 46.2%" in result.output
+    assert "actual: 46.4%" in result.output
     assert 1 == result.exit_code
 
 
@@ -41,29 +41,31 @@ def test_run_no_paths(runner, monkeypatch, tmpdir):
     "flags,exp_result,exp_exit_code",
     (
         # no flags
-        ([], 46.2, 1),
+        ([], 46.4, 1),
         # ignore init module
-        (["-I"], 46.0, 1),
+        (["-I"], 46.3, 1),
         # ignore module docs
-        (["-M"], 45.7, 1),
+        (["-M"], 46.0, 1),
         # ignore semiprivate docs
-        (["-s"], 46.7, 1),
+        (["-s"], 46.9, 1),
         # ignore private docs
-        (["-p"], 47.8, 1),
+        (["-p"], 48.0, 1),
+        # ignore property getter/setter decorators
+        (["-P"], 46.2, 1),
         # ignore magic method docs
-        (["-m"], 45.8, 1),
+        (["-m"], 46.2, 1),
         # ignore init method docs
-        (["-i"], 44.9, 1),
+        (["-i"], 45.3, 1),
         # ignore regex
-        (["-r", "^get$"], 45.8, 1),
+        (["-r", "^get$"], 46.2, 1),
         # whitelist regex
         (["-w", "^get$"], 50.0, 1),
         # exclude file
-        (["-e", os.path.join(SAMPLE_DIR, "partial.py")], 53.1, 1),
+        (["-e", os.path.join(SAMPLE_DIR, "partial.py")], 55.9, 1),
         # exclude file which doesn't exist
-        (["-e", os.path.join(SAMPLE_DIR, "does.not.exist")], 46.2, 1),
+        (["-e", os.path.join(SAMPLE_DIR, "does.not.exist")], 46.4, 1),
         # fail under
-        (["-f", "40"], 46.2, 0),
+        (["-f", "40"], 46.4, 0),
     ),
 )
 def test_run_shortflags(flags, exp_result, exp_exit_code, runner):
@@ -79,16 +81,17 @@ def test_run_shortflags(flags, exp_result, exp_exit_code, runner):
 @pytest.mark.parametrize(
     "flags,exp_result,exp_exit_code",
     (
-        (["--ignore-init-module"], 46.0, 1),
-        (["--ignore-module"], 45.7, 1),
-        (["--ignore-semiprivate"], 46.7, 1),
-        (["--ignore-private"], 47.8, 1),
-        (["--ignore-magic"], 45.8, 1),
-        (["--ignore-init-method"], 44.9, 1),
-        (["--ignore-regex", "^get$"], 45.8, 1),
+        (["--ignore-init-module"], 46.3, 1),
+        (["--ignore-module"], 46.0, 1),
+        (["--ignore-semiprivate"], 46.9, 1),
+        (["--ignore-private"], 48.0, 1),
+        (["--ignore-property-decorators"], 46.2, 1),
+        (["--ignore-magic"], 46.2, 1),
+        (["--ignore-init-method"], 45.3, 1),
+        (["--ignore-regex", "^get$"], 46.2, 1),
         (["--whitelist-regex", "^get$"], 50.0, 1),
-        (["--exclude", os.path.join(SAMPLE_DIR, "partial.py")], 53.1, 1),
-        (["--fail-under", "40"], 46.2, 0),
+        (["--exclude", os.path.join(SAMPLE_DIR, "partial.py")], 55.9, 1),
+        (["--fail-under", "40"], 46.4, 0),
     ),
 )
 def test_run_longflags(flags, exp_result, exp_exit_code, runner):
@@ -104,9 +107,9 @@ def test_run_longflags(flags, exp_result, exp_exit_code, runner):
 @pytest.mark.parametrize(
     "flags,exp_result,exp_exit_code",
     (
-        (["-i", "-I", "-r" "^method_foo$"], 45.5, 1),
-        (["-s", "-p", "-M"], 48.5, 1),
-        (["-m", "-f", "45"], 45.8, 0),
+        (["-i", "-I", "-r" "^method_foo$"], 45.8, 1),
+        (["-s", "-p", "-M"], 48.6, 1),
+        (["-m", "-f", "45"], 46.2, 0),
     ),
 )
 def test_run_multiple_flags(flags, exp_result, exp_exit_code, runner):